Issue #16 #17
Added ability for the server to broadcast the messages
as they are received to the clients. This needs to be checked
as it is possible the sockets are falling out of scope of the
handler before they can respond to the message sender.
This will be addressed in a future issue/update.